requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-extensions==4.0.0
attrs==23.2.0
Brotli==1.1.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dj-config-url==0.1.1
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==3.5.0
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-versatileimagefield==3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-flex-fields==0.8.6
entrypoint==2.0.3
exceptiongroup==1.2.1
funcs==0.10.1
gitdb==4.0.11
GitPython==3.1.43
gunicorn==22.0.0
h11==0.14.0
idna==3.7
importlib-metadata==2.0.0
iters==0.18.0
Markdown==3.3
mixed-methods==1.1.1
named==1.4.2
orderings==1.4.0
packaging==24.0
Pillow==8.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dotenv==1.0.1
python-gettext==5.0
python-magic==0.4.27
pytz==2020.1
requests==2.32.3
six==1.15.0
smmap==5.0.1
sniffio==1.3.1
solus==1.2.2
sqlparse==0.4.1
typing==3.7.4.3
typing-aliases==1.10.1
typing_extensions==4.12.0
urllib3==2.2.1
uvicorn==0.30.1
versions==2.1.2
whitenoise==6.6.0
wraps==0.12.0
zipp==3.3.0
